Splitting up graph
